DiscussionIndexResource tests * refactoring to move the resource api to djangosherd, where the DiscussionIndex model is located * adding tests
